refactoring (wip)
