+++ cluster/plugin/aws/plugin_test.go
@@ -1,0 +1,10 @@
+package aws
+
+import (
+	"log"
+	"testing"
+)
+
+func TestFoo(t *testing.T) {
+	log.Println("test...")
+}
